requirements.txt
tk
opencv-python
imutils
Pillow



